W2 - Bad Tutor
Satirical version of an essay tutor that subverts the role of a tutor, especially an AI tutor where students assume it'll respond to their question accurately and without hesitation. Instead of providing easy access to content, this tutor provides an alternative essay that writes with parallel formatting but on a completely different topic. This serves as actually a better scaffold and tutor model than the 'Good Tutor' that provides students with straight up essay guides for their essay question needs.